+++ connection_editor.cpp
@@ -12,34 +12,12 @@
  * this file belongs to.
  *****************************************************************************/
 #include "connection_editor.h"
-#include "logging.hpp"
+#include "bezier_editor_handle.h"
 
 #include <QBrush>
 #include <QGraphicsScene>
 #include <QGraphicsSceneMouseEvent>
 #include <QPainter>
-
-namespace {
-
-class BezierEditorHandle : public QGraphicsItem {
-    std::function<void(const QPointF& pos)> fn_;
-    QPointF src_pos_;
-    QPainterPath shape_;
-    QRectF ellipse_;
-
-public:
-    BezierEditorHandle(const QPoint& srcPoint, const QPoint& bezyOffset,
-        QGraphicsItem* parent, std::function<void(const QPointF& pos)> fn);
-
-    QRectF handleRect() const;
-    void setHandlePos(const QPointF& pos);
-
-    void mouseMoveEvent(QGraphicsSceneMouseEvent* event) final;
-    QRectF boundingRect() const final;
-    void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget = nullptr) final;
-};
-
-}
 
 namespace ceam {
 
@@ -72,7 +50,7 @@ void ConnectionEditor::setConnectionData(const ConnectionData& data)
                 data_.setBezyCtlPoint0(newPos - data_.sourcePoint());
                 emit connectionUpdated(data_);
             });
-        shape_.addEllipse(c0->handleRect());
+        shape_.addPath(c0->shape());
         handles_.append(c0);
 
         auto c1 = new BezierEditorHandle(
@@ -83,7 +61,7 @@ void ConnectionEditor::setConnectionData(const ConnectionData& data)
                 data_.setBezyCtlPoint1(newPos - data_.destinationPoint());
                 emit connectionUpdated(data_);
             });
-        shape_.addEllipse(c1->handleRect());
+        shape_.addPath(c1->shape());
         handles_.append(c1);
 
     } break;
@@ -116,54 +94,3 @@ void ConnectionEditor::paint(QPainter* painter, const QStyleOptionGraphicsItem* 
 }
 
 } // namespace ceam
-
-BezierEditorHandle::BezierEditorHandle(const QPoint& srcPoint, const QPoint& bezyOffset, QGraphicsItem* parent, std::function<void(const QPointF&)> fn)
-    : QGraphicsItem(parent)
-    , fn_(fn)
-    , src_pos_(srcPoint)
-    , ellipse_ { -5, -5, 10, 10 }
-{
-    setToolTip("Segment point");
-    setFlag(ItemIsMovable);
-    setHandlePos(src_pos_ + bezyOffset);
-}
-
-QRectF BezierEditorHandle::handleRect() const
-{
-    return ellipse_.translated(pos());
-}
-
-void BezierEditorHandle::setHandlePos(const QPointF& pos)
-{
-    prepareGeometryChange();
-    shape_.clear();
-    shape_.addEllipse(ellipse_);
-    shape_.addPolygon({ mapFromScene(src_pos_), mapFromScene(pos) });
-    setPos(pos);
-}
-
-void BezierEditorHandle::mouseMoveEvent(QGraphicsSceneMouseEvent* event)
-{
-    event->accept();
-    setHandlePos(event->scenePos());
-
-    if (fn_)
-        fn_(event->scenePos());
-
-    update(boundingRect());
-}
-
-QRectF BezierEditorHandle::boundingRect() const
-{
-    return shape_.boundingRect();
-}
-
-void BezierEditorHandle::paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget)
-{
-    painter->setPen(QPen(Qt::darkGray, 1));
-    painter->drawLine({ mapFromScene(src_pos_), mapFromScene(pos()) });
-
-    painter->setPen(QPen(Qt::darkGray, 1));
-    painter->setBrush(Qt::green);
-    painter->drawEllipse(ellipse_);
-}